r97: minor fixes for cygwin and sources

+++ libusb/os/windows_usb.c
@@ -157,7 +157,8 @@ static char* guid_to_string(const GUID guid)
 {
 static char guid_string[GUID_STRING_LENGTH];
 	
-	sprintf(guid_string, "{%08X-%04X-%04X-%02X%02X-%02X%02X%02X%02X%02X%02X}", guid.Data1, guid.Data2, guid.Data3, 
+	sprintf(guid_string, "{%08X-%04X-%04X-%02X%02X-%02X%02X%02X%02X%02X%02X}",
+		(unsigned int)guid.Data1, guid.Data2, guid.Data3, 
 		guid.Data4[0], guid.Data4[1], guid.Data4[2], guid.Data4[3],
 		guid.Data4[4], guid.Data4[5], guid.Data4[6], guid.Data4[7]);
 	return guid_string;
